+++ lib/core/constants/colors.dart
@@ -1,81 +1,109 @@
 import 'package:flutter/material.dart';
 
-abstract class PNDColors {
-  final Color transparent = Colors.transparent;
-  final Color red = const Color(0XFFFF0000);
+class AppColor extends ThemeExtension<AppColor> {
+  factory AppColor() => _light;
 
-  //Primary Color
-  final Color jnPrimary = const Color(0XFF0DCC5A);
-  final Color jnGreenText = const Color(0XFF0CB650);
-  final Color navigationBarColor =
-      const Color(0xFFFFFFFF); //const Color(0xFFf6f6f6);
+  AppColor._({
+    required this.white,
+    required this.black,
+    required this.blue,
+    required this.red,
+    required this.primaryGreen,
+    required this.yellowGreen,
+    required this.lightGreen,
+    required this.gray90,
+    required this.gray50,
+    required this.gray30,
+    required this.gray20,
+    required this.gray10,
+  });
 
-  //Status Color
-  final Color error = const Color(0xFFDA1E28);
-  final Color success = const Color(0xFF1B8EE4);
+  static final AppColor _light = AppColor._(
+    white: const Color(0xFFFFFFFF),
+    black: const Color(0xFF09090B),
+    blue: const Color(0xFF6A9DFF),
+    red: const Color(0xFFFF2727),
+    primaryGreen: const Color(0xFF10C200),
+    yellowGreen: const Color(0xFFB5E5A6),
+    lightGreen: const Color(0xFFE2F7E0),
+    gray90: const Color(0xFF333333),
+    gray50: const Color(0xFFACACAC),
+    gray30: const Color(0xFFD9D9D9),
+    gray20: const Color(0xFFF3F3F3),
+    gray10: const Color(0xFFF9F9F9),
+  );
 
-  //Grayscale Color
-  /// like black
-  final Color gray900 = const Color(0xFF141313); //230309
-  final Color gray800 = const Color(0xFF363C45);
-  final Color gray700 = const Color(0xFF5A616B);
-  final Color gray600 = const Color(0xFF787E89);
-  final Color gray500 = const Color(0xFF9CA3AF); //0722변경
-  final Color gray400 = const Color(0xFFC2C6CE); //0722변경
-  final Color gray300 = const Color(0xFFDADEE5); //0722변경
-  final Color gray200 = const Color(0xFFF1F4F6); //0722변경
-  final Color gray100 = const Color(0xFFF7F9FA);
+  final Color white;
+  final Color black;
+  final Color blue;
+  final Color red;
+  final Color primaryGreen;
+  final Color yellowGreen;
+  final Color lightGreen;
+  final Color gray90;
+  final Color gray50;
+  final Color gray30;
+  final Color gray20;
+  final Color gray10;
 
-  /// like white
-  final Color gray000 = const Color(0xFFFFFFFF);
+  static late BuildContext _context;
+  static void init(BuildContext context) => _context = context;
 
-  final Color naver = const Color(0xFF3BC229);
+  static AppColor get of => Theme.of(_context).extension<AppColor>()!;
+  static AppColor? get maybeOf => Theme.of(_context).extension<AppColor>();
 
-  final Color secondary00 = const Color(0XFF0057B7);
-  final Color secondary01 = const Color(0XFFFFB600);
+  @override
+  ThemeExtension<AppColor> copyWith({
+    Color? white,
+    Color? black,
+    Color? blue,
+    Color? red,
+    Color? primaryGreen,
+    Color? yellowGreen,
+    Color? lightGreen,
+    Color? gray90,
+    Color? gray50,
+    Color? gray30,
+    Color? gray20,
+    Color? gray10,
+  }) {
+    return AppColor._(
+      white: white ?? this.white,
+      black: black ?? this.black,
+      blue: blue ?? this.blue,
+      red: red ?? this.red,
+      primaryGreen: primaryGreen ?? this.primaryGreen,
+      yellowGreen: yellowGreen ?? this.yellowGreen,
+      lightGreen: lightGreen ?? this.lightGreen,
+      gray90: gray90 ?? this.gray90,
+      gray50: gray50 ?? this.gray50,
+      gray30: gray30 ?? this.gray30,
+      gray20: gray20 ?? this.gray20,
+      gray10: gray10 ?? this.gray10,
+    );
+  }
 
-  final Color accent00 = const Color(0XFF24A148);
-
-  final Color color_FFB800 = const Color(0XFFFFB800);
-
-  final Color payco_primary = const Color(0XFFE72A2D);
-  final Color naver_primary = const Color(0XFF00C73C);
-  final Color facebook_primary = const Color(0XFF375691);
-  final Color kakao_primary = const Color(0XFFFEE500);
-
-  final Color product_complete_noti = const Color(0XFFF9F9F9);
-
-  // 리브랜등 추가
-  final Color jnSub1 = const Color(0xFFFF5453);
-  final Color jnSub2 = const Color(0xFFFFF5F4);
-
-  //10% ~ 95%
-  final Color neutrals_gray_10 = const Color(0xffeaecf5);
-  final Color neutrals_gray_20 = const Color(0xffd6d7e4);
-  final Color neutrals_gray_30 = const Color(0xffafb1c0);
-  final Color neutrals_gray_40 = const Color(0xff8d8da1);
-  final Color neutrals_gray_50 = const Color(0xff67677f);
-  final Color neutrals_gray_60 = const Color(0xff444456);
-  final Color neutrals_gray_70 = const Color(0xff31313c);
-  final Color neutrals_gray_80 = const Color(0xff202029);
-  final Color neutrals_gray_90 = const Color(0xff191923);
-  final Color neutrals_gray_95 = const Color(0xff12121c);
-
-  final Color color_80000000 = const Color(0X80000000);
-  final Color color_CC000000 = const Color(0XCC000000);
-  final Color color_33000000 = const Color(0X33000000);
-  final Color color_FF000000 = const Color(0XFF000000);
-
-  final Color color_FFF6EC = const Color(0Xfffff6ec);
-
-  final Color color_FFF3EB = const Color(0Xfffff3eb);
-
-  final Color color_47c482 = const Color(0XFF47C482);
-  final Color color_0a0400 = const Color(0XFF0a0400);
-  final Color color_FFF6F0 = const Color(0XFFFFF6F0);
-  final Color color_F1F1F1 = const Color(0XFFF1F1F1);
-  final Color color_F3F3F3 = const Color(0XFFF3F3F3);
-  final Color color_C2C6CE = const Color(0XFFC2C6CE);
-  final Color color_CCF4DC = const Color(0XFFCCF4DC);
-  final Color color_019FB1 = const Color(0xFF019FB1);
+  @override
+  ThemeExtension<AppColor> lerp(
+    covariant ThemeExtension<AppColor>? other,
+    double t,
+  ) {
+    if (other is! AppColor) {
+      return this;
+    }
+    return AppColor._(
+      white: Color.lerp(white, other.white, t)!,
+      black: Color.lerp(black, other.black, t)!,
+      blue: Color.lerp(black, other.black, t)!,
+      red: Color.lerp(red, other.red, t)!,
+      primaryGreen: Color.lerp(primaryGreen, other.primaryGreen, t)!,
+      yellowGreen: Color.lerp(yellowGreen, other.yellowGreen, t)!,
+      lightGreen: Color.lerp(lightGreen, other.lightGreen, t)!,
+      gray90: Color.lerp(gray90, other.gray90, t)!,
+      gray50: Color.lerp(gray50, other.gray50, t)!,
+      gray30: Color.lerp(gray30, other.gray30, t)!,
+      gray20: Color.lerp(gray20, other.gray20, t)!,
+      gray10: Color.lerp(gray10, other.gray10, t)!,
+    );
+  }
 }
