--- lib/core/constants/font_weight.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PNDFontWeights {
-  static const FontWeight light = FontWeight.w200;
-
-  static const FontWeight normal = FontWeight.w400;
-  static const FontWeight fontWeight600 = FontWeight.w600;
-
-  static const FontWeight bold = FontWeight.bold;
-
-  static const FontWeight regular = FontWeight.w400;
-  static const FontWeight semiBold = FontWeight.w600;
-}
