+++ lib/core/constants/images.dart
@@ -1,4 +1,7 @@
-class PNDSvgs {
+class PNDImages {
+  static const String cat = 'assets/imgs/pet/cat.png';
+  static const String dog = 'assets/imgs/pet/dog.png';
+
   static const String mainIcon = 'assets/svgs/main_icon.svg';
   static const String mainTitleIcon = 'assets/svgs/main_title_icon.svg';
   static const String google = 'assets/svgs/login/google_icon.svg';
@@ -10,6 +13,6 @@ class PNDSvgs {
   static const String chat = 'assets/svgs/home/icon_chat.svg';
   static const String user = 'assets/svgs/home/icon_user.svg';
 
-  static const String cat = 'assets/svgs/pet/cat_active.svg';
-  static const String dog = 'assets/svgs/pet/dog_active.svg';
+  static const String catActive = 'assets/svgs/pet/cat_active.svg';
+  static const String dogActive = 'assets/svgs/pet/dog_active.svg';
 }
