--- lib/core/constants/imgs.dart
@@ -1,4 +1,0 @@
-class PNDImgs {
-  static const String cat = 'assets/imgs/pet/cat.png';
-  static const String dog = 'assets/imgs/pet/dog.png';
-}
