+++ lib/core/enums/home_home_tab_type.dart
@@ -1,0 +1,8 @@
+enum HomeHomeTabType {
+  petSos('돌봄급구'),
+  petMate('돌봄메이트');
+
+  const HomeHomeTabType(this.label);
+
+  final String label;
+}
