+++ lib/core/enums/sns_provider_type.dart
@@ -1,5 +1,3 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
 enum SnsProviderType {
   kakao(providerId: 'kakao.com'),
   google(providerId: 'google.com'),
@@ -19,24 +17,4 @@ enum SnsProviderType {
   }
 
   final String providerId;
-}
-
-enum PetType {
-  @JsonValue('cat')
-  cat(displayName: '고양이', code: 'cat'),
-  @JsonValue('dog')
-  dog(displayName: '강아지', code: 'dog');
-
-  const PetType({required this.displayName, required this.code});
-  final String displayName;
-  final String code;
-}
-
-enum HomeHomeTabType {
-  petSos('돌봄급구'),
-  petMate('돌봄메이트');
-
-  const HomeHomeTabType(this.label);
-
-  final String label;
 }
