+++ lib/features/auth/data/data_sources/sns_data_sources/kakao_auth_data_source_impl.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
-import 'package:pets_next_door_flutter/core/constants/enums.dart';
+import 'package:pets_next_door_flutter/core/enums/sns_provider_type.dart';
 import 'package:pets_next_door_flutter/features/auth/data/data_sources/sns_data_sources/sns_auth_data_source.dart';
 import 'package:pets_next_door_flutter/features/auth/domain/sns_oauth_info.dart';
