+++ lib/features/pet/domain/breed.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:pets_next_door_flutter/core/constants/enums.dart';
+import 'package:pets_next_door_flutter/core/enums/pet_type.dart';
 
 part 'breed.g.dart';
 part 'breed.freezed.dart';
