+++ lib/presentation/pages/sign_in/widgets/sns_button_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:pets_next_door_flutter/core/constants/enums.dart';
-import 'package:pets_next_door_flutter/core/constants/svgs.dart';
+import 'package:pets_next_door_flutter/core/enums/sns_provider_type.dart';
+import 'package:pets_next_door_flutter/core/constants/images.dart';
 
 /// 터치할 수 있는 원형 sns 버튼 위젯
 /// - [onTap] : 선택한 sns provider type을 리턴합니다.
@@ -20,9 +20,9 @@ class SnsButtonWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final svgAssetUrl = switch (snsType) {
-      SnsProviderType.apple => PNDSvgs.apple,
-      SnsProviderType.google => PNDSvgs.google,
-      SnsProviderType.kakao => PNDSvgs.kakao,
+      SnsProviderType.apple => PNDImages.apple,
+      SnsProviderType.google => PNDImages.google,
+      SnsProviderType.kakao => PNDImages.kakao,
     };
 
     return GestureDetector(
