+++ lib/presentation/pages/sign_in/widgets/start_with_apple_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:pets_next_door_flutter/core/constants/svgs.dart';
+import 'package:pets_next_door_flutter/core/constants/images.dart';
 import 'package:pets_next_door_flutter/presentation/pages/sign_in/sign_in_event.dart';
 
 /// 터치할 수 있는 원형 sns 버튼 위젯
@@ -16,7 +16,7 @@ class StartWithAppleButton extends ConsumerWidget with SignInEvent {
       onTap: () async {
         await onTapSignInWithApple(ref);
       },
-      child: SvgPicture.asset(PNDSvgs.apple),
+      child: SvgPicture.asset(PNDImages.apple),
     );
   }
 }
