+++ lib/presentation/pages/splash/splash_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pets_next_door_flutter/core/theme/extension/app_color.dart';
+import 'package:pets_next_door_flutter/core/constants/colors.dart';
 
 class SplashPage extends StatelessWidget {
   const SplashPage({super.key});
